Moving 'core' files into mgt-core folder
