+++ src/bundle/index.es6.ts
@@ -5,6 +5,6 @@
  * -------------------------------------------------------------------------------------------
  */
 
+export * from '../mgt-core';
 export * from '../index';
-export * from '../components/baseComponent';
 export * from '../mock/mgt-mock-provider';
