+++ src/components/mgt-agenda/mgt-agenda.ts
@@ -7,13 +7,10 @@
 
 import * as MicrosoftGraph from '@microsoft/microsoft-graph-types';
 import { customElement, html, property, TemplateResult } from 'lit-element';
-import { Providers } from '../../Providers';
-import { ProviderState } from '../../providers/IProvider';
+import { getDayOfWeekString, getMonthString, MgtTemplatedComponent, Providers, ProviderState } from '../../mgt-core';
 import '../../styles/fabric-icon-font';
 import { prepScopes } from '../../utils/GraphHelpers';
-import { getDayOfWeekString, getMonthString } from '../../utils/Utils';
 import '../mgt-person/mgt-person';
-import { MgtTemplatedComponent } from '../templatedComponent';
 import { styles } from './mgt-agenda-css';
 import { getEvents } from './mgt-agenda.graph';
