+++ src/components/mgt-get/mgt-get.ts
@@ -6,11 +6,8 @@
  */
 
 import { customElement, html, property } from 'lit-element';
-import { Providers } from '../../Providers';
-import { ProviderState } from '../../providers/IProvider';
+import { equals, MgtTemplatedComponent, Providers, ProviderState } from '../../mgt-core';
 import { prepScopes } from '../../utils/GraphHelpers';
-import { equals } from '../../utils/Utils';
-import { MgtTemplatedComponent } from '../templatedComponent';
 
 /**
  * Custom element for making Microsoft Graph get queries
