+++ src/components/mgt-login/mgt-login.ts
@@ -7,10 +7,8 @@
 
 import { User } from '@microsoft/microsoft-graph-types';
 import { customElement, html, property } from 'lit-element';
-import { Providers } from '../../Providers';
-import { ProviderState } from '../../providers/IProvider';
+import { MgtBaseComponent, Providers, ProviderState } from '../../mgt-core';
 import '../../styles/fabric-icon-font';
-import { MgtBaseComponent } from '../baseComponent';
 import '../mgt-person/mgt-person';
 import { styles } from './mgt-login-css';
