+++ src/components/mgt-tasks/mgt-tasks.graph.planner.ts
@@ -6,7 +6,7 @@
  */
 
 import { PlannerBucket, PlannerPlan, PlannerTask } from '@microsoft/microsoft-graph-types';
-import { IGraph } from '../../IGraph';
+import { IGraph } from '../../mgt-core';
 import { prepScopes } from '../../utils/GraphHelpers';
 
 /**
