+++ src/components/providers/mgt-proxy-provider.ts
@@ -5,10 +5,9 @@
  * -------------------------------------------------------------------------------------------
  */
 
-import { customElement, LitElement, property } from 'lit-element';
-import { Providers } from '../../Providers';
+import { customElement, property } from 'lit-element';
+import { MgtBaseProvider, Providers } from '../../mgt-core';
 import { ProxyProvider } from '../../providers/ProxyProvider';
-import { MgtBaseProvider } from './baseProvider';
 
 /**
  * Authentication component for ProxyProvider
