+++ src/components/sub-components/mgt-dot-options/mgt-dot-options.ts
@@ -7,8 +7,9 @@
 
 import { customElement, html, property } from 'lit-element';
 import { classMap } from 'lit-html/directives/class-map';
-import { MgtBaseComponent } from '../../baseComponent';
+import { MgtBaseComponent } from '../../../mgt-core';
 import { styles } from './mgt-dot-options-css';
+
 /**
  * Custom Component used to handle an arrow rendering for TaskGroups utilized in the task component.
  *
