+++ src/graph/graph.user.ts
@@ -6,7 +6,7 @@
  */
 
 import { User } from '@microsoft/microsoft-graph-types';
-import { IGraph } from '../IGraph';
+import { IGraph } from '../mgt-core';
 import { prepScopes } from '../utils/GraphHelpers';
 
 /**
