+++ src/index.ts
@@ -7,6 +7,4 @@
 
 export * from './components/components';
 export * from './components/providers/providers';
-export * from './Providers';
 export * from './providers/providers';
-export { prepScopes } from './utils/GraphHelpers';
