+++ src/mgt-core/Providers.ts
@@ -7,6 +7,7 @@
 
 import { IProvider } from './providers/IProvider';
 import { EventDispatcher, EventHandler } from './utils/EventDispatcher';
+
 /**
  * Provides implementation for acquiring the necessary access token for calling the Microsoft Graph APIs.
  *
