+++ src/providers/WamProvider.ts
@@ -6,8 +6,8 @@
  */
 
 import { AuthenticationProviderOptions } from '@microsoft/microsoft-graph-client/lib/es/IAuthenticationProviderOptions';
-import { createFromProvider, Graph } from '../Graph';
-import { IProvider, ProviderState } from './IProvider';
+import { createFromProvider } from '../Graph';
+import { IProvider, ProviderState } from '../mgt-core';
 
 /**
  * Windows auth
