+++ src/providers/providers.ts
@@ -5,7 +5,6 @@
  * -------------------------------------------------------------------------------------------
  */
 
-export * from './IProvider';
 export * from './MsalProvider';
 export * from './SharePointProvider';
 export * from './TeamsProvider';
